% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comperes-package.R
\docType{package}
\name{comperes-package}
\alias{comperes}
\alias{comperes-package}
\title{comperes: Create and Manipulate Competition Results}
\description{
Tools for creating and manipulating competition results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/echasnovski/comperes}
  \item Report bugs at \url{https://github.com/echasnovski/comperes/issues}
}

}
\author{
\strong{Maintainer}: Evgeni Chasnovski \email{evgeni.chasnovski@gmail.com}

}
