% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head-to-head-new.R
\name{h2h-long}
\alias{h2h-long}
\alias{h2h_long}
\alias{to_h2h_long}
\title{Compute long format of Head-to-Head values}
\usage{
h2h_long(cr_data, ..., fill = list())

to_h2h_long(mat, value = "h2h_value", drop = FALSE)
}
\arguments{
\item{cr_data}{Competition results ready for \code{\link[=as_longcr]{as_longcr()}}.}

\item{...}{Name-value pairs of Head-to-Head functions (see Details).}

\item{fill}{A named list that for each variable supplies a single value to
use instead of NA for missing pairs (see tidyr's
\link[tidyr:complete]{complete()}).}

\item{mat}{Matrix of Head-to-Head values.}

\item{value}{String name to be used for column with Head-to-Head value.}

\item{drop}{Use \code{TRUE} to drop rows with missing Head-to-Head values (see
Details).}
}
\value{
An object of class \code{h2h_long} which is a
\link[tibble:tibble]{tibble} with columns \code{player1}, \code{player2} and those,
produced by Head-to-Head functions (for \code{h2h_long()} maybe none).
}
\description{
Functions to compute Head-to-Head values in long pair-value format.
}
\details{
\code{h2h_long()} computes Head-to-Head values in
\link[=convert-pair-value]{long} format. It allows computation of multiple
Head-to-Head values at the same time by supplying multiple summary
functions in \code{...}. If no function is supplied in \code{...}

After computing Head-to-Head values of present matchups, implicit missing
matchups are turned into explicit (by adding corresponding rows with filling
values in Head-to-Head columns) by using tidyr's
\link[tidyr:complete]{complete()}. Use \code{fill} as in \code{complete()} to control
filling values. To drop those rows use tidyr's \link[tidyr:drop_na]{drop_na()}.

\code{to_h2h_long()} takes object of \link[=h2h-mat]{h2h_mat} structure and converts it
into \code{h2h_long} object with value column named as stored in \code{value}. Use
\code{drop = TRUE} to remove rows with missing values in value column (but not in
players').
}
\section{Head-to-Head value}{

Head-to-Head value is a summary statistic of players'
\link[=get_cr_matchups]{matchups}. In other words, every game is converted in
series of "subgames" between ordered pairs of players (including selfplay)
which is stored as \link[=results-widecr]{widecr} object. After that summary of
item, defined by columns \code{player1} and \code{player2} is computed using
\code{\link[=summarise_item]{summarise_item()}}.

That said, name-value pairs of Head-to-Head functions should be defined as
for \code{summarise_item()} applied to data with columns \code{game}, \code{player1},
\code{score1}, \code{player2}, \code{score2}.
}

\examples{
ncaa2005 \%>\%
  h2h_long(
    mean_score = mean(score1),
    mean_abs_score = mean(abs(score1 - score2))
  )

ncaa2005[-(1:2), ] \%>\%
  h2h_long(
    mean_score = mean(score1),
    fill = list(mean_score = 0)
  )

}
