% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{levels2}
\alias{levels2}
\title{Levels of vector}
\usage{
levels2(x, na.last = TRUE)
}
\arguments{
\item{x}{An object of interest.}

\item{na.last}{Argument for controlling the treatment of \code{NA}s. See \code{\link[=sort]{sort()}}.}
}
\description{
Extension of \code{\link[=levels]{levels()}} function. If \code{levels(x)} is not \code{NULL}, it is
returned. Otherwise, character representation of unique sorted values is
returned (with \code{NA} treated based on \code{na.last} as in \code{\link[=sort]{sort()}}).
}
\examples{
fac_vec <- factor(c("a", "b"), levels = c("a", "b", "c"))
levels2(fac_vec)

levels2(c(10, 1, 2, NA, 11))

}
