% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head-to-head-new.R
\name{num_wins}
\alias{num_wins}
\title{Compute number of wins}
\usage{
num_wins(player1, score1, player2, score2, half_for_draw)
}
\arguments{
\item{player1}{Vector of \code{player1} identifiers.}

\item{score1}{Vector of scores for \code{player1}.}

\item{player2}{Vector of \code{player2} identifiers.}

\item{score2}{Vector of scores for \code{player2}.}

\item{half_for_draw}{Use \code{TRUE} to add half the mathups with draws.}
}
\description{
Function to accompany \code{num_wins} and \code{num_wins2} in \link{h2h_funs}. May be useful
for outer usage.
}
\keyword{internal}
