% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tidy-eval.R
\name{tidyeval}
\alias{tidyeval}
\alias{quo}
\alias{syms}
\title{Tidy eval helpers}
\description{
These functions provide tidy eval-compatible ways to capture symbols
(\code{syms()}) and quosures (\code{quo()}). To learn more about tidy eval and how to
use these tools, read
\url{http://rlang.tidyverse.org/articles/tidy-evaluation.html}
}
\keyword{internal}
