% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head-to-head.R
\name{num_wins}
\alias{num_wins}
\title{Compute number of wins}
\usage{
num_wins(score_1, score_2, half_for_draw = FALSE, na.rm = TRUE)
}
\arguments{
\item{score_1}{Vector of scores for first player.}

\item{score_2}{Vector of scores for second player.}

\item{half_for_draw}{Use \code{TRUE} to add half the matchups with draws.}

\item{na.rm}{Use \code{TRUE} to not count pair of scores with at least one \code{NA}.}
}
\description{
Function to accompany \code{num_wins} and \code{num_wins2} in \link{h2h_funs}. May be useful
for outer usage.
}
\keyword{internal}
