% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tidy-eval.R
\name{tidyeval}
\alias{tidyeval}
\alias{quo}
\alias{enquo}
\alias{sym}
\alias{syms}
\alias{expr}
\alias{enexpr}
\title{Tidy eval helpers}
\description{
These functions provide tidy eval-compatible ways to capture symbols
(\code{sym()}, \code{syms()}), expressions (\code{expr()}, \code{enexpr()}), and quosures
(\code{quo()}, \code{enquo()}).
}
\keyword{internal}
